Fix and disable MSVC warnings
Needed because some warnings are no longer disabled after
http://crrev.com/c/3189512.

Also includes https://github.com/KhronosGroup/OpenCL-Headers/pull/179,
needed after clang upgrade to llvmorg-14-init-5410-gd0473681

Bug: chromium:1257173
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3210629
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>

+++ src/tests/gl_tests/ReadPixelsTest.cpp
@@ -274,7 +274,7 @@ TEST_P(ReadPixelsPBONVTest, ReadFromFBOWithDataOffset)
 
     // Read (height - 1) rows offset by width * 4.
     glReadPixels(0, 0, mFBOWidth, mFBOHeight - 1, GL_RGBA, GL_UNSIGNED_BYTE,
-                 reinterpret_cast<void *>(mFBOWidth * 4));
+                 reinterpret_cast<void *>(mFBOWidth * static_cast<uintptr_t>(4)));
 
     void *mappedPtr =
         glMapBufferRangeEXT(GL_PIXEL_PACK_BUFFER, 0, 4 * mFBOWidth * mFBOHeight, GL_MAP_READ_BIT);
